% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{beta_draw}
\alias{beta_draw}
\title{Draw beta from N(omegabar * x_invsigma_y, omegabar)}
\usage{
beta_draw(invomega0, x_invsigma_x, x_invsigma_y, random_vec, symb)
}
\description{
Draw beta from N(omegabar * x_invsigma_y, omegabar)
}
