% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{beta_symbolic}
\alias{beta_symbolic}
\title{Perform a symbolic decomposition of invomega0 + x_invsigma_x and return a pointer to it}
\usage{
beta_symbolic(invomega0, x_invsigma_x)
}
\description{
Perform a symbolic decomposition of invomega0 + x_invsigma_x and return a pointer to it
}
